@@ -1,15 +1,17 @@
 {
   "name": "elm-pages",
-  "version": "2.1.11",
+  "type": "module",
+  "version": "3.0.0-beta.42",
   "homepage": "https://elm-pages.com",
   "moduleResolution": "node",
   "description": "Type-safe static sites, written in pure elm with your own custom elm-markup syntax.",
   "main": "index.js",
   "scripts": {
-    "start": "cd examples/end-to-end && npm start",
-    "test": "npx elmi-to-json --version && elm-verify-examples --run-tests && elm-test && (cd generator && mocha) && (cd examples/routing && npm i && npm run build -- --debug && elm-test) && npm run test:snapshot",
-    "test:snapshot": "(cd examples/escaping && npm install && npm test && cd ../..) && (cd examples/base-path && npm install && npm test && cd ../..)",
-    "cypress": "npm start & cypress run --spec cypress/integration/elm-pages-dev.spec.js",
+    "start": "cd examples/end-to-end && npm i && npx elm-pages dev",
+    "test": "./test.sh",
+    "test:snapshot": "(cd examples/escaping && npm install && npm test) && (cd examples/base-path && npm install && npm test)",
+    "cypress": "npm start & cypress run",
+    "build:generator": "elm-codegen install && (cd codegen && lamdera make Generate.elm --output elm-pages-codegen.js && mv elm-pages-codegen.js elm-pages-codegen.cjs)",
     "review": "elm-review"
   },
   "repository": "https://github.com/dillonkearns/elm-pages",
@@ -23,45 +25,55 @@
   "author": "Dillon Kearns",
   "license": "BSD-3-Clause",
   "dependencies": {
-    "chokidar": "3.5.2",
-    "commander": "8.3.0",
+    "busboy": "^1.6.0",
+    "chokidar": "^3.5.3",
+    "commander": "^10.0.1",
     "connect": "^3.7.0",
+    "cookie-signature": "^1.2.1",
     "cross-spawn": "7.0.3",
-    "devcert": "^1.2.0",
+    "devcert": "^1.2.2",
     "elm-doc-preview": "^5.0.5",
     "elm-hot": "^1.1.6",
-    "fs-extra": "^10.0.0",
-    "globby": "11.0.4",
+    "esbuild": "^0.17.18",
+    "fs-extra": "^11.1.1",
+    "globby": "13.1.4",
     "gray-matter": "^4.0.3",
     "jsesc": "^3.0.2",
-    "kleur": "^4.1.4",
-    "micromatch": "^4.0.4",
-    "object-hash": "^2.2.0",
-    "serve-static": "^1.14.1",
-    "terser": "^5.7.2",
-    "undici": "^4.4.7",
-    "which": "^2.0.2"
+    "kleur": "^4.1.5",
+    "make-fetch-happen": "^11.1.1",
+    "memfs": "^3.5.1",
+    "micromatch": "^4.0.5",
+    "serve-static": "^1.15.0",
+    "terser": "^5.17.1",
+    "vite": "^4.3.5",
+    "which": "^3.0.1"
   },
   "devDependencies": {
     "@types/cross-spawn": "^6.0.2",
-    "@types/fs-extra": "9.0.12",
+    "@types/fs-extra": "^11.0.1",
+    "@types/make-fetch-happen": "^10.0.1",
     "@types/micromatch": "^4.0.2",
-    "@types/node": "12.20.12",
-    "@types/serve-static": "1.13.10",
-    "cypress": "^8.3.0",
-    "elm-optimize-level-2": "^0.1.5",
-    "elm-review": "^2.5.3",
-    "elm-test": "^0.19.1-revision7",
-    "elm-tooling": "^1.3.0",
-    "elm-verify-examples": "^5.0.0",
-    "elmi-to-json": "^1.2.0",
-    "mocha": "^9.1.0",
-    "typescript": "4.3.5"
+    "@types/node": "^20.1.0",
+    "@types/serve-static": "^1.15.1",
+    "cypress": "^12.13.0",
+    "elm-codegen": "^0.3.0",
+    "elm-optimize-level-2": "^0.3.5",
+    "elm-review": "^2.10.2",
+    "elm-test": "^0.19.1-revision12",
+    "elm-tooling": "^1.14.0",
+    "elm-verify-examples": "^5.2.0",
+    "elmi-to-json": "^1.4.3",
+    "typescript": "^5.0.4",
+    "vite": "^4.3.5",
+    "vitest": "^0.31.0"
   },
   "files": [
+    "adapter/",
     "generator/src/",
     "generator/review/",
+    "generator/dead-code-review/",
     "src/",
+    "codegen/elm-pages-codegen.cjs",
     "generator/template/",
     "generator/static-code/"
   ],
