@@ -1,36 +1,47 @@
-const fs = require("./dir-helpers.js");
-const fsPromises = require("fs").promises;
+import * as fs from "./dir-helpers.js";
+import * as fsPromises from "fs/promises";
+import { runElmReview } from "./compile-elm.js";
+import { restoreColorSafe } from "./error-formatter.js";
+import * as path from "path";
+import { spawn as spawnCallback } from "cross-spawn";
+import * as codegen from "./codegen.js";
+import * as terser from "terser";
+import * as os from "os";
+import { Worker, SHARE_ENV } from "worker_threads";
+import { ensureDirSync } from "./file-helpers.js";
+import { generateClientFolder } from "./codegen.js";
+import { default as which } from "which";
+import { build } from "vite";
+import * as preRenderHtml from "./pre-render-html.js";
+import * as esbuild from "esbuild";
+import { createHash } from "crypto";
+import { merge_vite_configs } from "./vite-utils.js";
+import { resolveConfig } from "./config.js";
+import * as globby from "globby";
+import { fileURLToPath } from "url";
+import { copyFile } from "fs/promises";
 
-const { runElmReview } = require("./compile-elm.js");
-const { restoreColor } = require("./error-formatter");
-const path = require("path");
-const spawnCallback = require("cross-spawn").spawn;
-const codegen = require("./codegen.js");
-const terser = require("terser");
-const os = require("os");
-const { Worker, SHARE_ENV } = require("worker_threads");
-const { ensureDirSync } = require("./file-helpers.js");
-const which = require("which");
 let pool = [];
 let pagesReady;
+let pagesErrored;
 let pages = new Promise((resolve, reject) => {
   pagesReady = resolve;
+  pagesErrored = reject;
 });
+let pagesReadyCalled = false;
+let activeWorkers = 0;
 let buildError = false;
 
-const DIR_PATH = path.join(process.cwd());
 const OUTPUT_FILE_NAME = "elm.js";
 
 process.on("unhandledRejection", (error) => {
-  console.trace("Unhandled: ", error);
+  console.log(error);
   process.exitCode = 1;
 });
 
-const ELM_FILE_PATH = path.join(
-  DIR_PATH,
-  "./elm-stuff/elm-pages",
-  OUTPUT_FILE_NAME
-);
+function ELM_FILE_PATH() {
+  return path.join(process.cwd(), "./elm-stuff/elm-pages", OUTPUT_FILE_NAME);
+}
 
 async function ensureRequiredDirs() {
   ensureDirSync(`dist`);
@@ -39,9 +50,9 @@ async function ensureRequiredDirs() {
 
 async function ensureRequiredExecutables() {
   try {
-    await which("elm");
+    await which("lamdera");
   } catch (error) {
-    throw "I couldn't find elm on the PATH. Please ensure it's installed, either globally, or locally. If it's installed locally, ensure you're running through an NPM script or with npx so the PATH is configured to include it.";
+    throw "I couldn't find lamdera on the PATH. Please ensure it's installed, either globally, or locally. If it's installed locally, ensure you're running through an NPM script or with npx so the PATH is configured to include it.";
   }
   try {
     await which("elm-optimize-level-2");
@@ -55,51 +66,251 @@ async function ensureRequiredExecutables() {
   }
 }
 
-async function run(options) {
+export async function run(options) {
+  const __filename = fileURLToPath(import.meta.url);
+  const __dirname = path.dirname(__filename);
+  console.warn = function (...messages) {
+    // This is a temporary hack to avoid this warning. elm-pages manages compiling the Elm code without Vite's involvement, so it is external to Vite.
+    // There is a pending issue to allow having external scripts in Vite, once this issue is fixed we can remove this hack:
+    // https://github.com/vitejs/vite/issues/3533
+    if (!messages[0]?.startsWith(`<script src="/elm.js">`)) {
+      console.info(...messages);
+    }
+  };
   try {
     await ensureRequiredDirs();
     await ensureRequiredExecutables();
-    // since init/update are never called in pre-renders, and DataSource.Http is called using undici
+    // since init/update are never called in pre-renders, and BackendTask.Http is called using pure NodeJS HTTP fetching
     // we can provide a fake HTTP instead of xhr2 (which is otherwise needed for Elm HTTP requests from Node)
-    XMLHttpRequest = {};
 
     const generateCode = codegen.generate(options.base);
 
-    const copyDone = copyAssets();
     await generateCode;
 
-    const compileCli = compileCliApp(options);
-    const compileClientDone = compileElm(options);
-    await compileCli;
-    const cliDone = runCli(options);
+    const config = await resolveConfig();
+    await fsPromises.writeFile(
+      "elm-stuff/elm-pages/index.html",
+      preRenderHtml.templateHtml(false, config.headTagsTemplate)
+    );
+    const viteConfig = merge_vite_configs(
+      {
+        configFile: false,
+        root: process.cwd(),
+        base: options.base,
+        ssr: false,
 
-    await Promise.all([copyDone, cliDone, compileClientDone]);
-  } catch (error) {
-    buildError = true;
+        build: {
+          manifest: true,
+          outDir: "dist",
+          rollupOptions: {
+            input: "elm-stuff/elm-pages/index.html",
+          },
+        },
+        optimizeDeps: {
+          include: [],
+        },
+      },
+      config.vite || {}
+    );
+
+    const buildComplete = build(viteConfig);
+    const compileClientDone = compileElm(options);
+    await buildComplete;
+    await compileClientDone;
+    const fullOutputPath = path.join(process.cwd(), `./dist/elm.js`);
+    const withoutExtension = path.join(process.cwd(), `./dist/elm`);
+    const browserElmHash = await fingerprintElmAsset(
+      fullOutputPath,
+      withoutExtension
+    );
+    const assetManifestPath = path.join(process.cwd(), "dist/manifest.json");
+    const manifest = JSON.parse(
+      await fsPromises.readFile(assetManifestPath, { encoding: "utf-8" })
+    );
+    const indexTemplate = await fsPromises.readFile(
+      "dist/elm-stuff/elm-pages/index.html",
+      "utf-8"
+    );
+    const preloadFiles = [
+      `elm.${browserElmHash}.js`,
+      ...Object.entries(manifest).map((entry) => entry[1].file),
+    ].map((file) => path.join(options.base, file));
+    const userProcessedPreloads = preloadFiles.flatMap((file) => {
+      const userPreloadForFile = config.preloadTagForFile(file);
+      if (userPreloadForFile === true) {
+        return [defaultPreloadForFile(file)];
+      } else if (userPreloadForFile === false) {
+        return [];
+      } else if (typeof userPreloadForFile === "string") {
+        return [userPreloadForFile];
+      } else {
+        throw `I expected preloadTagForFile in elm-pages.config.mjs to return a string or boolean, but instead it returned: ${userPreloadForFile}`;
+      }
+    });
+
+    const processedIndexTemplate = indexTemplate
+      .replace("<!-- PLACEHOLDER_PRELOADS -->", userProcessedPreloads.join(""))
+      .replace(
+        '<script defer src="/elm.js" type="text/javascript"></script>',
+        `<script defer src="/elm.${browserElmHash}.js" type="text/javascript"></script>`
+      );
+    await fsPromises.writeFile("dist/template.html", processedIndexTemplate);
+    await fsPromises.unlink(assetManifestPath);
+    const portBackendTaskCompiled = esbuild
+      .build({
+        entryPoints: ["./custom-backend-task"],
+        platform: "node",
+        outfile: ".elm-pages/compiled-ports/custom-backend-task.mjs",
+        assetNames: "[name]-[hash]",
+        chunkNames: "chunks/[name]-[hash]",
+        outExtension: { ".js": ".js" },
+        metafile: true,
+        bundle: true,
+        format: "esm",
+        packages: "external",
+        logLevel: "silent",
+      })
+      .then((result) => {
+        try {
+          global.portsFilePath = Object.keys(result.metafile.outputs)[0];
+        } catch (e) {}
+      })
+      .catch((error) => {
+        const portBackendTaskFileFound =
+          globby.globbySync("./custom-backend-task.*").length > 0;
+        if (portBackendTaskFileFound) {
+          // don't present error if there are no files matching custom-backend-task
+          // if there are files matching custom-backend-task, warn the user in case something went wrong loading it
+          console.error("Failed to start custom-backend-task watcher", error);
+        }
+      });
+
+    global.XMLHttpRequest = {};
+    const compileCli = compileCliApp(options);
     try {
+      await compileCli;
+      await compileClientDone;
+      await portBackendTaskCompiled;
+      const inlineRenderCode = `
+import * as renderer from "./render.js";
+import * as elmModule from "${path.resolve("./elm-stuff/elm-pages/elm.cjs")}";
+import * as url from 'url';
+${
+  global.portsFilePath
+    ? `import * as customBackendTask from "${path.resolve(
+        global.portsFilePath
+      )}";`
+    : `const customBackendTask = {};`
+}
+
+import * as preRenderHtml from "./pre-render-html.js";
+const basePath = \`${options.base || "/"}\`;
+const htmlTemplate = ${JSON.stringify(processedIndexTemplate)};
+const mode = "build";
+const addWatcher = () => {};
+
+export async function render(request) {
+  const requestTime = new Date();
+  const response = await renderer.render(
+    customBackendTask,
+    basePath,
+    elmModule.default,
+    mode,
+    (new url.URL(request.rawUrl)).pathname,
+    request,
+    addWatcher,
+    false
+  );
+  console.dir(response);
+  if (response.kind === "bytes") {
+    return {
+        body: response.contentDatPayload.buffer,
+        statusCode: response.statusCode,
+        kind: response.kind,
+        headers: response.headers,
+    }
+  } else if (response.kind === "api-response") {
+    // isBase64Encoded
+    return {
+        body: response.body.body,
+        statusCode: response.statusCode,
+        kind: response.kind,
+        headers: response.headers,
+        isBase64Encoded: response.body.isBase64Encoded,
+    }
+  } else {
+    return {
+        body: preRenderHtml.replaceTemplate(htmlTemplate, response.htmlString),
+        statusCode: response.statusCode,
+        kind: response.kind,
+        headers: response.headers,
+    }
+  }
+}
+`;
+      await esbuild.build({
+        format: "esm",
+        platform: "node",
+        stdin: { contents: inlineRenderCode, resolveDir: __dirname },
+        bundle: true,
+        // TODO do I need to make the outfile joined with the current working directory?
+
+        outfile: ".elm-pages/compiled/render.mjs",
+        // external: ["node:*", ...options.external],
+        packages: "external",
+        minify: true,
+        // absWorkingDir: projectDirectory,
+        // banner: { js: `#!/usr/bin/env node\n\n${ESM_REQUIRE_SHIM}` },
+      });
+    } catch (cliError) {
+      // TODO make sure not to print duplicate error output if cleaner review output is printed
+      console.error(cliError);
       const reviewOutput = JSON.parse(await runElmReview());
       const isParsingError = reviewOutput.errors.some((reviewError) => {
         return reviewError.errors.some((item) => item.rule === "ParsingError");
       });
       if (isParsingError) {
-        console.error(error);
+        console.error(cliError);
       } else {
-        console.error(restoreColor(reviewOutput));
+        console.error(restoreColorSafe(reviewOutput));
       }
       process.exit(1);
-    } catch (noElmReviewErrors) {
-      console.error(error);
-    } finally {
-      process.exit(1);
     }
+    await portBackendTaskCompiled;
+    const cliDone = runCli(options);
+    await cliDone;
+
+    await fsPromises.rename(
+      "dist/____elm-pages-internal____/404/index.html",
+      "dist/404.html"
+    );
+    await runAdapter(
+      config.adapter ||
+        function () {
+          console.log(
+            "No adapter configured in elm-pages.config.mjs. Skipping adapter step."
+          );
+        },
+      processedIndexTemplate
+    );
+  } catch (error) {
+    console.trace(error);
+    if (error) {
+      console.error(restoreColorSafe(error));
+    }
+    buildError = true;
+    process.exitCode = 1;
   }
 }
 
 /**
  * @param {string} basePath
  */
-function initWorker(basePath) {
+function initWorker(basePath, whenDone) {
   return new Promise((resolve, reject) => {
+    const __filename = fileURLToPath(import.meta.url);
+    const __dirname = path.dirname(__filename);
+    activeWorkers += 1;
     let newWorker = {
       worker: new Worker(path.join(__dirname, "./render-worker.js"), {
         env: SHARE_ENV,
@@ -109,21 +320,32 @@ function initWorker(basePath) {
     newWorker.worker.once("online", () => {
       newWorker.worker.on("message", (message) => {
         if (message.tag === "all-paths") {
-          pagesReady(JSON.parse(message.data));
+          pagesReadyCalled = true;
+          pagesReady(
+            ["/____elm-pages-internal____/404"].concat(JSON.parse(message.data))
+          );
         } else if (message.tag === "error") {
           process.exitCode = 1;
-          console.error(restoreColor(message.data));
-          buildNextPage(newWorker);
+          console.error(restoreColorSafe(message.data));
+          if (!pagesReadyCalled) {
+            // when there is a build error while resolving all-paths, we don't know which pages to build so we need to short-circuit
+            // and give an error instead of trying to build the remaining pages to show as many errors as possible
+            pagesReady([]);
+            reject(message.data);
+          }
+          buildError = true;
+          buildNextPage(newWorker, whenDone);
         } else if (message.tag === "done") {
-          buildNextPage(newWorker);
+          buildNextPage(newWorker, whenDone);
         } else {
           throw `Unhandled tag ${message.tag}`;
         }
       });
       newWorker.worker.on("error", (error) => {
         console.error("Unhandled worker exception", error);
+        buildError = true;
         process.exitCode = 1;
-        buildNextPage(newWorker);
+        buildNextPage(newWorker, whenDone);
       });
       resolve(newWorker);
     });
@@ -134,57 +356,94 @@ function initWorker(basePath) {
  */
 function prepareStaticPathsNew(thread) {
   thread.worker.postMessage({
+    portsFilePath: global.portsFilePath,
     mode: "build",
     tag: "render",
     pathname: "/all-paths.json",
   });
 }
 
-async function buildNextPage(thread) {
+async function buildNextPage(thread, allComplete) {
   let nextPage = (await pages).pop();
   if (nextPage) {
     thread.worker.postMessage({
+      portsFilePath: global.portsFilePath,
       mode: "build",
       tag: "render",
       pathname: nextPage,
     });
   } else {
     thread.worker.terminate();
+    activeWorkers -= 1;
+    allComplete();
   }
 }
 
-async function runCli(options) {
-  const cpuCount = os.cpus().length;
-  console.log("Threads: ", cpuCount);
+function runCli(options) {
+  return new Promise((resolve, reject) => {
+    const whenDone = () => {
+      if (activeWorkers === 0) {
+        // wait for the remaining tasks in the pool to complete once the pages queue is emptied
+        Promise.all(pool).then((value) => {
+          if (buildError) {
+            reject();
+          } else {
+            resolve(value);
+          }
+        });
+      }
+    };
+    const cpuCount = os.cpus().length;
+    // const cpuCount = 1;
+    console.log("Threads: ", cpuCount);
 
-  const getPathsWorker = initWorker(options.base);
-  getPathsWorker.then(prepareStaticPathsNew);
-  const threadsToCreate = Math.max(1, cpuCount / 2 - 1);
-  pool.push(getPathsWorker);
-  for (let index = 0; index < threadsToCreate - 1; index++) {
-    pool.push(initWorker(options.base));
-  }
-  pool.forEach((threadPromise) => {
-    threadPromise.then(buildNextPage);
+    const getPathsWorker = initWorker(options.base, whenDone);
+    getPathsWorker.then(prepareStaticPathsNew);
+    const threadsToCreate = Math.max(1, cpuCount - 1);
+    pool.push(getPathsWorker);
+    for (let index = 0; index < threadsToCreate - 1; index++) {
+      pool.push(initWorker(options.base, whenDone));
+    }
+    pool.forEach((threadPromise) => {
+      threadPromise.then((thread) => buildNextPage(thread, whenDone));
+    });
   });
 }
 
 async function compileElm(options) {
-  const outputPath = `dist/elm.js`;
-  const fullOutputPath = path.join(process.cwd(), `dist/elm.js`);
-  await spawnElmMake(options, ".elm-pages/TemplateModulesBeta.elm", outputPath);
+  ensureDirSync("dist");
+  const fullOutputPath = path.join(process.cwd(), `./dist/elm.js`);
+  await generateClientFolder(options.base);
+  await spawnElmMake(
+    options.debug ? "debug" : "optimize",
+    options,
+    ".elm-pages/Main.elm",
+    fullOutputPath,
+    path.join(process.cwd(), "./elm-stuff/elm-pages/client")
+  );
 
   if (!options.debug) {
     await runTerser(fullOutputPath);
   }
 }
 
-function elmOptimizeLevel2(elmEntrypointPath, outputPath, cwd) {
+async function fingerprintElmAsset(fullOutputPath, withoutExtension) {
+  const fileHash = await fsPromises
+    .readFile(fullOutputPath, "utf8")
+    .then(getAssetHash);
+  await fsPromises.copyFile(
+    fullOutputPath,
+    `${withoutExtension}.${fileHash}.js`
+  );
+  return fileHash;
+}
+
+export function elmOptimizeLevel2(outputPath, cwd) {
   return new Promise((resolve, reject) => {
-    const fullOutputPath = cwd ? path.join(cwd, outputPath) : outputPath;
+    const optimizedOutputPath = outputPath + ".opt";
     const subprocess = spawnCallback(
       `elm-optimize-level-2`,
-      [elmEntrypointPath, "--output", outputPath],
+      [outputPath, "--output", optimizedOutputPath],
       {
         // ignore stdout
         // stdio: ["inherit", "ignore", "inherit"],
@@ -199,17 +458,16 @@ function elmOptimizeLevel2(elmEntrypointPath, outputPath, cwd) {
     });
 
     subprocess.on("close", async (code) => {
-      if (
-        code === 0 &&
-        commandOutput === "" &&
-        (await fs.fileExists(fullOutputPath))
-      ) {
+      if (code === 0) {
+        await copyFile(optimizedOutputPath, outputPath);
         resolve();
       } else {
         if (!buildError) {
           buildError = true;
           process.exitCode = 1;
-          reject(commandOutput);
+          reject(
+            `I encountered an error when running elm-optimize-level-2:\n\n ${commandOutput}`
+          );
         } else {
           // avoid unhandled error printing duplicate message, let process.exit in top loop take over
         }
@@ -218,29 +476,67 @@ function elmOptimizeLevel2(elmEntrypointPath, outputPath, cwd) {
   });
 }
 
+/** @typedef {"debug" | "optimize" | "default"} CompileMode  */
+
 /**
+ * @param {CompileMode} mode
  * @param {string} elmEntrypointPath
  * @param {string} outputPath
  * @param {string | undefined} cwd
  */
-async function spawnElmMake(options, elmEntrypointPath, outputPath, cwd) {
-  if (options.debug) {
-    await runElmMake(options, elmEntrypointPath, outputPath, cwd);
+async function spawnElmMake(mode, options, elmEntrypointPath, outputPath, cwd) {
+  await runElmMake(mode, options, elmEntrypointPath, outputPath, cwd);
+  if (mode === "optimize") {
+    await elmOptimizeLevel2(outputPath, cwd);
+  }
+  await fsPromises.writeFile(
+    outputPath,
+    (
+      await fsPromises.readFile(outputPath, "utf-8")
+    ).replace(
+      /return \$elm\$json\$Json\$Encode\$string\(.REPLACE_ME_WITH_FORM_TO_STRING.\)/g,
+      "function appendSubmitter (myFormData, event) { event.submitter && event.submitter.name && event.submitter.name.length > 0 ? myFormData.append(event.submitter.name, event.submitter.value) : myFormData;  return myFormData }; return " +
+        (options.debug
+          ? "_Json_wrap(Array.from(appendSubmitter(new FormData(_Json_unwrap(event).target), _Json_unwrap(event))))"
+          : "[...(appendSubmitter(new FormData(event.target), event))]")
+    )
+  );
+}
+
+function getAssetHash(content) {
+  return createHash("sha256").update(content).digest("hex").slice(0, 8);
+}
+
+/**
+ * @param {CompileMode} mode
+ */
+function modeToOptions(mode) {
+  if (mode === "debug") {
+    return ["--debug"];
+  } else if (mode === "optimize") {
+    return ["--optimize"];
   } else {
-    await elmOptimizeLevel2(elmEntrypointPath, outputPath, cwd);
+    return [];
   }
 }
 
-function runElmMake(options, elmEntrypointPath, outputPath, cwd) {
+/**
+ * @param {CompileMode} mode
+ * @param {string} elmEntrypointPath
+ * @param {string} outputPath
+ * @param {string | undefined} cwd
+ */
+function runElmMake(mode, options, elmEntrypointPath, outputPath, cwd) {
   return new Promise(async (resolve, reject) => {
+    const executableName = options.executableName || "lamdera";
     const subprocess = spawnCallback(
-      `elm`,
+      executableName,
       [
         "make",
         elmEntrypointPath,
         "--output",
         outputPath,
-        ...(options.debug ? ["--debug"] : []),
+        ...modeToOptions(mode),
         "--report",
         "json",
       ],
@@ -251,9 +547,8 @@ function runElmMake(options, elmEntrypointPath, outputPath, cwd) {
         cwd: cwd,
       }
     );
-    const fullOutputPath = cwd ? path.join(cwd, outputPath) : outputPath;
-    if (await fs.fileExists(fullOutputPath)) {
-      await fsPromises.unlink(fullOutputPath, {
+    if (await fs.fileExists(outputPath)) {
+      await fsPromises.unlink(outputPath, {
         force: true /* ignore errors if file doesn't exist */,
       });
     }
@@ -262,15 +557,25 @@ function runElmMake(options, elmEntrypointPath, outputPath, cwd) {
     subprocess.stderr.on("data", function (data) {
       commandOutput += data;
     });
+    subprocess.on("error", function () {
+      reject(commandOutput);
+    });
 
     subprocess.on("close", async (code) => {
-      if (code == 0 && (await fs.fileExists(fullOutputPath))) {
+      if (
+        code == 0 &&
+        (await fs.fileExists(outputPath)) &&
+        commandOutput === ""
+      ) {
         resolve();
       } else {
         if (!buildError) {
           buildError = true;
-          process.exitCode = 1;
-          reject(restoreColor(JSON.parse(commandOutput)));
+          try {
+            reject(restoreColorSafe(commandOutput));
+          } catch (error) {
+            reject(commandOutput);
+          }
         } else {
           // avoid unhandled error printing duplicate message, let process.exit in top loop take over
         }
@@ -282,7 +587,7 @@ function runElmMake(options, elmEntrypointPath, outputPath, cwd) {
 /**
  * @param {string} filePath
  */
-async function runTerser(filePath) {
+export async function runTerser(filePath) {
   console.log("Running terser");
   const minifiedElm = await terser.minify(
     (await fsPromises.readFile(filePath)).toString(),
@@ -325,19 +630,17 @@ async function runTerser(filePath) {
   }
 }
 
-async function copyAssets() {
-  fs.copyDirFlat("public", "dist");
-}
-
-async function compileCliApp(options) {
+export async function compileCliApp(options) {
   await spawnElmMake(
+    // TODO should be --optimize, but there seems to be an issue with the html to JSON with --optimize
+    options.debug ? "debug" : "optimize",
     options,
-    ".elm-pages/TemplateModulesBeta.elm",
-    "elm.js",
-    "./elm-stuff/elm-pages"
+    path.join(process.cwd(), "elm-stuff/elm-pages/.elm-pages/Main.elm"),
+    path.join(process.cwd(), "elm-stuff/elm-pages/elm.js"),
+    path.join(process.cwd(), "elm-stuff/elm-pages")
   );
 
-  const elmFileContent = await fsPromises.readFile(ELM_FILE_PATH, "utf-8");
+  const elmFileContent = await fsPromises.readFile(ELM_FILE_PATH(), "utf-8");
   // Source: https://github.com/elm-explorations/test/blob/d5eb84809de0f8bbf50303efd26889092c800609/src/Elm/Kernel/HtmlAsJson.js
   const forceThunksSource = ` _HtmlAsJson_toJson(x)
 }
@@ -354,7 +657,9 @@ async function compileCliApp(options) {
   }
 
 function forceThunks(vNode) {
-  if (typeof vNode !== "undefined" && vNode.$ === "#2") {
+if ( (typeof vNode !== "undefined" && vNode.$ === "#2") // normal/debug mode
+     || (typeof vNode !== "undefined" && typeof vNode.$ === "undefined" && typeof vNode.a == "string" && typeof vNode.b == "object" ) // optimize mode
+   ) {
     // This is a tuple (the kids : List (String, Html) field of a Keyed node); recurse into the right side of the tuple
     vNode.b = forceThunks(vNode.b);
   }
@@ -381,31 +686,64 @@ function _HtmlAsJson_toJson(html) {
 `;
 
   await fsPromises.writeFile(
-    ELM_FILE_PATH,
+    ELM_FILE_PATH().replace(/\.js$/, ".cjs"),
     elmFileContent
       .replace(
         /return \$elm\$json\$Json\$Encode\$string\(.REPLACE_ME_WITH_JSON_STRINGIFY.\)/g,
-        "return " +
-          (options.debug
-            ? `${forceThunksSource}
+        `return ${forceThunksSource}
   return _Json_wrap(forceThunks(html));
 `
-            : `${forceThunksSource}
-return forceThunks(html);
-`)
       )
-      .replace(
-        "return ports ? { ports: ports } : {};",
-        `const die = function() {
-        managers = null
-        model = null
-        stepper = null
-        ports = null
-      }
-
-      return ports ? { ports: ports, die: die } : { die: die };`
-      )
+      .replace(`console.log('App dying')`, "")
   );
+}
+
+async function runAdapter(adaptFn, processedIndexTemplate) {
+  try {
+    await adaptFn({
+      renderFunctionFilePath: "./.elm-pages/compiled/render.mjs",
+      routePatterns: JSON.parse(
+        await fsPromises.readFile("./dist/route-patterns.json", "utf-8")
+      ),
+      apiRoutePatterns: JSON.parse(
+        await fsPromises.readFile("./dist/api-patterns.json", "utf-8")
+      ),
+    });
+    console.log("Success - Adapter script complete");
+  } catch (error) {
+    console.trace("ERROR - Adapter script failed", error);
+    try {
+      console.error(JSON.stringify(error));
+    } catch (parsingError) {
+      console.error(error);
+    }
+    process.exit(1);
+  }
+}
+
+// Source: https://github.com/vitejs/vite/blob/c53ffec3465d2d28d08d29ca61313469e03f5dd6/playground/ssr-vue/src/entry-server.js#L50-L68
+/**
+ * @param {string} file
+ */
+function defaultPreloadForFile(file) {
+  if (file.endsWith(".js")) {
+    return `<link rel="modulepreload" crossorigin href="${file}">`;
+  } else if (file.endsWith(".css")) {
+    return `<link rel="preload" href="${file}" as="style">`;
+  } else if (file.endsWith(".woff")) {
+    return ` <link rel="preload" href="${file}" as="font" type="font/woff" crossorigin>`;
+  } else if (file.endsWith(".woff2")) {
+    return ` <link rel="preload" href="${file}" as="font" type="font/woff2" crossorigin>`;
+  } else if (file.endsWith(".gif")) {
+    return ` <link rel="preload" href="${file}" as="image" type="image/gif">`;
+  } else if (file.endsWith(".jpg") || file.endsWith(".jpeg")) {
+    return ` <link rel="preload" href="${file}" as="image" type="image/jpeg">`;
+  } else if (file.endsWith(".png")) {
+    return ` <link rel="preload" href="${file}" as="image" type="image/png">`;
+  } else {
+    // TODO
+    return "";
+  }
 }
 
 /** @typedef { { route : string; contentJson : string; head : SeoTag[]; html: string; body: string; } } FromElm */
@@ -422,5 +760,3 @@ return forceThunks(html);
  * @param {string} contentJsonString
  * @returns {string}
  */
-
-module.exports = { run };
