@@ -1,116 +1,81 @@
-const cliVersion = require("../../package.json").version;
-const seo = require("./seo-renderer.js");
-const elmPagesJsMinified = require("./elm-pages-js-minified.js");
-const path = require("path").posix;
-const jsesc = require("jsesc");
+import * as seo from "./seo-renderer.js";
+import * as path from "node:path";
+import { fileURLToPath } from "node:url";
+import { packageVersion } from "./compatibility-key.js";
 
-/** @typedef { { head: any[]; errors: any[]; contentJson: any[]; html: string; route: string; title: string; } } Arg */
+/** @typedef { { head: any[]; errors: any[]; html: string; route: string; title: string; } } Arg */
 /** @typedef { { tag : 'PageProgress'; args : Arg[] } } PageProgress */
 
-module.exports =
-  /**
-   * @param {string} basePath
-   * @param {Arg} fromElm
-   * @param {unknown} contentJson
-   * @param {boolean} devServer
-   * @returns {string}
-   */
-  function wrapHtml(basePath, fromElm, contentJson, devServer) {
-    const devServerOnly = (/** @type {string} */ devServerOnlyString) =>
-      devServer ? devServerOnlyString : "";
-    const seoData = seo.gather(fromElm.head);
-    /*html*/
-    return `<!DOCTYPE html>
-  ${seoData.rootElement}
-  <head>
-    <link rel="stylesheet" href="${path.join(basePath, "style.css")}">
-    ${devServerOnly(devServerStyleTag())}
-    <link rel="preload" href="${path.join(basePath, "elm.js")}" as="script">
-    <link rel="modulepreload" href="${path.join(basePath, "index.js")}">
-    ${devServerOnly(
-      /* html */ `<script defer="defer" src="${path.join(
-      basePath,
-      "hmr.js"
-    )}" type="text/javascript"></script>`
-    )}
-    <script defer="defer" src="${path.join(
-      basePath,
-      "elm.js"
-    )}" type="text/javascript"></script>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width,initial-scale=1">
-    <script type="module">
-import userInit from"${path.join(basePath, "index.js")}";
-${elmPagesJsMinified}
-    </script>
-    <title>${fromElm.title}</title>
-    <meta name="generator" content="elm-pages v${cliVersion}">
-    <link rel="manifest" href="${path.join(basePath, "manifest.json")}">
-    <meta name="mobile-web-app-capable" content="yes">
-    <meta name="theme-color" content="#ffffff">
-    <meta name="apple-mobile-web-app-capable" content="yes">
-    <meta name="apple-mobile-web-app-status-bar-style" content="black-translucent">
-    ${seoData.headTags}
-    <script id="__ELM_PAGES_DATA__" type="application/json">${jsesc(
-      contentJson,
-      {
-        isScriptContext: true,
-        json: true,
-      }
-    )}</script>
-    </head>
-    <body>
-      <div data-url="" display="none"></div>
-      ${fromElm.html}
-    </body>
-  </html>
-  `;
+export function wrapHtml(basePath, fromElm, contentDatPayload) {
+  const seoData = seo.gather(fromElm.head);
+  return {
+    kind: "html-template",
+    title: fromElm.title,
+    html: fromElm.html,
+    bytesData: Buffer.from(contentDatPayload.buffer).toString("base64"),
+    headTags: seoData.headTags,
+    rootElement: seoData.rootElement,
   };
-
-/**
- * @param {string} route
- */
-function cleanRoute(route) {
-  return route.replace(/(^\/|\/$)/, "");
 }
 
 /**
- * @param {string} cleanedRoute
+ * @param {boolean} devMode
+ * @param {(context: {cliVersion: string;}) => string} userHeadTagsTemplate
  */
-function pathToRoot(cleanedRoute) {
-  return cleanedRoute === ""
-    ? cleanedRoute
-    : cleanedRoute
-      .split("/")
-      .map((_) => "..")
-      .join("/")
-      .replace(/\.$/, "./");
-}
-
-function devServerStyleTag() {
-  /*html*/
-  return `<style>
-@keyframes lds-default {
-    0%, 20%, 80%, 100% {
-      transform: scale(1);
+export function templateHtml(devMode, userHeadTagsTemplate) {
+  const __filename = fileURLToPath(import.meta.url);
+  const __dirname = path.dirname(__filename);
+  return /* html */ `<!DOCTYPE html>
+<!-- ROOT --><html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <title><!-- PLACEHOLDER_TITLE --></title>
+    ${
+      devMode
+        ? `<script src="/hmr.js" type="text/javascript"></script>
+    <link rel="stylesheet" href="/dev-style.css">`
+        : `<!-- PLACEHOLDER_PRELOADS -->`
     }
-    50% {
-      transform: scale(1.5);
+    <script defer src="/elm.js" type="text/javascript"></script>
+    ${
+      devMode
+        ? `<script src="/elm-pages.js" type="module"></script>`
+        : `<script defer src="${path.join(
+            __dirname,
+            "../static-code/elm-pages.js"
+          )}" type="module"></script>`
     }
-  }
-
-#not-found-reason code {
-  color: rgb(226, 0, 124);
+    ${indent(userHeadTagsTemplate({ cliVersion: packageVersion }))}
+    <!-- PLACEHOLDER_HEAD_AND_DATA -->
+  </head>
+  <body>
+    <div data-url="" display="none"></div>
+    <!-- PLACEHOLDER_HTML -->
+  </body>
+</html>`;
 }
 
-#not-found-reason h1 {
-  font-size: 26px;
-  font-weight: bold;
-  padding-bottom: 15px;
+/**
+ * @param {string} snippet
+ */
+function indent(snippet) {
+  return snippet
+    .split("\n")
+    .map((line) => `    ${line}`)
+    .join("\n");
 }
 
-#not-found-reason a:hover {
-  text-decoration: underline;
-}
-</style>`;
+/**
+ * @param {string} processedTemplate
+ */
+export function replaceTemplate(processedTemplate, info) {
+  return processedTemplate
+    .replace(
+      /<!--\s*PLACEHOLDER_HEAD_AND_DATA\s*-->/,
+      `${info.headTags}
+                  <script id="__ELM_PAGES_BYTES_DATA__" type="application/octet-stream">${info.bytesData}</script>`
+    )
+    .replace(/<!--\s*PLACEHOLDER_TITLE\s*-->/, info.title)
+    .replace(/<!--\s*PLACEHOLDER_HTML\s* -->/, info.html)
+    .replace(/<!-- ROOT -->\S*<html lang="en">/m, info.rootElement);
 }
