@@ -1,8 +1,7 @@
-import userInit from "/index.js";
+import userInit from "/index";
 
-let prefetchedPages = [window.location.pathname];
-let initialLocationHash = document.location.hash.replace(/^#/, "");
-
+let prefetchedPages;
+let initialLocationHash;
 /**
  * @returns
  */
@@ -11,15 +10,15 @@ function loadContentAndInitializeApp() {
   if (!path.endsWith("/")) {
     path = path + "/";
   }
-  const app = Elm.TemplateModulesBeta.init({
+  const app = Elm.Main.init({
     flags: {
       secrets: null,
       isPrerendering: false,
       isDevServer: false,
       isElmDebugMode: false,
-      contentJson: JSON.parse(
-        document.getElementById("__ELM_PAGES_DATA__").innerHTML
-      ),
+      contentJson: {},
+      pageDataBase64: document.getElementById("__ELM_PAGES_BYTES_DATA__")
+        .innerHTML,
       userFlags: userInit.flags(),
     },
   });
@@ -48,48 +47,51 @@ function prefetchIfNeeded(/** @type {HTMLAnchorElement} */ target) {
     link.setAttribute("as", "fetch");
 
     link.setAttribute("rel", "prefetch");
-    link.setAttribute("href", origin + target.pathname + "/content.json");
+    link.setAttribute("href", origin + target.pathname + "/content.dat");
     document.head.appendChild(link);
   }
 }
 
-const appPromise = new Promise(function (resolve, reject) {
-  document.addEventListener("DOMContentLoaded", (_) => {
-    resolve(loadContentAndInitializeApp());
-  });
-});
-
-userInit.load(appPromise);
-
-if (typeof connect === "function") {
-  connect(function (newContentJson) {
-    appPromise.then((app) => {
-      app.ports.fromJsPort.send({ contentJson: newContentJson });
+export function setup() {
+  prefetchedPages = [window.location.pathname];
+  initialLocationHash = document.location.hash.replace(/^#/, "");
+  const appPromise = new Promise(function (resolve, reject) {
+    document.addEventListener("DOMContentLoaded", (_) => {
+      resolve(loadContentAndInitializeApp());
     });
   });
-}
+  userInit.load(appPromise);
 
-/** @param {MouseEvent} event */
-const trigger_prefetch = (event) => {
-  const a = find_anchor(/** @type {Node} */ (event.target));
-  if (a && a.href && a.hasAttribute("elm-pages:prefetch")) {
-    prefetchIfNeeded(a);
+  if (typeof connect === "function") {
+    connect(function (bytesData) {
+      appPromise.then((app) => {
+        app.ports.hotReloadData.send(bytesData);
+      });
+    });
   }
-};
 
-/** @type {NodeJS.Timeout} */
-let mousemove_timeout;
+  /** @param {MouseEvent} event */
+  const trigger_prefetch = (event) => {
+    const a = find_anchor(/** @type {Node} */ (event.target));
+    if (a && a.href && a.hasAttribute("elm-pages:prefetch")) {
+      prefetchIfNeeded(a);
+    }
+  };
 
-/** @param {MouseEvent} event */
-const handle_mousemove = (event) => {
-  clearTimeout(mousemove_timeout);
-  mousemove_timeout = setTimeout(() => {
-    trigger_prefetch(event);
-  }, 20);
-};
+  /** @type {NodeJS.Timeout} */
+  let mousemove_timeout;
 
-addEventListener("touchstart", trigger_prefetch);
-addEventListener("mousemove", handle_mousemove);
+  /** @param {MouseEvent} event */
+  const handle_mousemove = (event) => {
+    clearTimeout(mousemove_timeout);
+    mousemove_timeout = setTimeout(() => {
+      trigger_prefetch(event);
+    }, 20);
+  };
+
+  addEventListener("touchstart", trigger_prefetch);
+  addEventListener("mousemove", handle_mousemove);
+}
 
 /**
  * @param {Node} node
@@ -100,3 +102,15 @@ function find_anchor(node) {
   while (node && node.nodeName.toUpperCase() !== "A") node = node.parentNode; // SVG <a> elements have a lowercase name
   return /** @type {HTMLAnchorElement} */ (node);
 }
+
+Object.defineProperty(SubmitEvent.prototype, "fields", {
+  get: function fields() {
+    let formData = new FormData(this.currentTarget);
+    if (this.submitter?.name) {
+      formData.append(this.submitter.name, this.submitter.value);
+    }
+    return [...formData.entries()];
+  },
+});
+
+setup();
